requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
yfinance==0.2.33
pandas-ta==0.3.14b0
openai==1.6.1
google-generativeai==0.3.2
anthropic==0.8.1
python-dotenv==1.0.0
plotly==5.18.0
apscheduler==3.10.4
pytz==2023.3
